Persist the NodeExternalIP using basic lookup
Motivation
Till we have the QEMU guest agent support for RHCOS we must report the
known external IP, otherwise there will be no linkage between the node
and machine object.

Modification
The external IP of a node is the IPv4 address of the node name. Every
node published using mdns publisher its name and address, in the cluster
network so just back resolve this.

Result
The machine object has the addresses value set correctly with with
its NodeExternalIP set to the resolvable IP of its name.

Related-to: https://bugzilla.redhat.com/show_bug.cgi?id=1764804

Signed-off-by: Roy Golan <rgolan@redhat.com>