Initiate the ovirt api connection only on demand
At first the kube api client is not ready, so we can't pull the secrets.
First the controller needs to initiate caches and then we can start work
with it.
The solution is initiate the ovirt client on the first usage or in case
of an error, to overcome connection close or token expirations.